requirements.txt
flask
requests
flask-socketio
simple-websocket
pyaudio
cryptography
pydub
pypqc